ServiceAPITests を 分解して 基底クラスRequestTestBase と継承されたOriginServiceAPIExample の追加
